// src/three-dimensional/tube/models/tube-definition.ts
export interface TubeDefinition {
  outerRadius: number;
  innerRadius: number;
  height: number;
  outerCircumference: number;
  innerCircumference: number;
  ExternalLateralSurfaceArea: number;
  InternalLateralSurfaceArea: number;
  outerArea: number;
  innerArea: number;
  endSurfaceArea: number;
  outerVolume: number;
  innerVolume: number;
  solidVolume: number;
  wallThickness: number;
}